chore(ui-scripts): refactor publish script to bypass lerna and publish directly to npm
Closes: INSTUI-2990
