+++ .github/workflows/manual-legacy-release-to-npm.yml
@@ -1,4 +1,4 @@
-name: Release current version to npm
+name: Release older version to npm
 on: workflow_dispatch
 jobs:
   release:
@@ -19,17 +19,20 @@ jobs:
           NPM_TOKEN: ${{secrets.NPM_TOKEN}}
           NPM_EMAIL: ${{secrets.NPM_EMAIL}}
           NPM_USERNAME: ${{secrets.NPM_USERNAME}}
-        run: yarn release maintenance
-      - name: Get commit message
-        run: |
-          echo "commmit_message=$(git log --format=%B -n 1 ${{ github.event.after }})" >> $GITHUB_ENV
+        run: yarn release current maintenance
+  tag:
+    needs: release
+    if: "startsWith(github.event.head_commit.message, 'chore(release)')"
+    runs-on: ubuntu-latest
+    name: Tag release commit
+    steps:
+      - uses: actions/checkout@v2
+        with:
+          fetch-depth: 0
       - name: Set up git identity
-        if: "startsWith(env.commmit_message, 'chore(release)')"
         run: git config --global user.name "instructure-ui-ci" && git config --global user.email "instructure-ui-ci@instructure.com"
         # -a -m is needed to make the tag "annotated". Lerna needs annotated tags to work correctly with verisons
       - name: Add tag
-        if: "startsWith(env.commmit_message, 'chore(release)')"
         run: git tag -a v$(./.github/workflows/calculateVersion.sh) -m v$(./.github/workflows/calculateVersion.sh)
       - name: Push tags
-        if: "startsWith(env.commmit_message, 'chore(release)')"
         run: git push origin v$(./.github/workflows/calculateVersion.sh)
